Collapse anytree_based/ subpackage into a single anytree_based.py
The legacy anytree-based tree implementation is ~170 lines of tightly
coupled code. Splitting it across 6 files (base_node, action_node,
belief_node, visualization, optimal_action, __init__) added friction
without benefit: lazy imports inside ActionNode.print / BeliefNode.print
to break the visualization cycle, forward-reference re-imports at the
bottom of action_node.py and belief_node.py, more files to navigate.

Consolidating to a single anytree_based.py mirrors arena.py's shape
(one tree implementation per file). The lazy imports and noqa pragmas
go away because everything is in module scope at call time.

Top-level re-exports in core/tree/__init__.py are unchanged, so
existing 'from POMDPPlanners.core.tree import ActionNode' callers
continue to work without modification.
