--- POMDPPlanners/core/tree/anytree_based/__init__.py
@@ -1,31 +1,0 @@
-"""Anytree-based implementation of the MCTS tree.
-
-Each node is a Python object (subclass of :class:`anytree.NodeMixin`)
-with attribute references to its parent and children. This is the
-original tree implementation, preserved alongside the column-store
-``arena`` implementation in :mod:`POMDPPlanners.core.tree.arena`.
-
-Use::
-
-    from POMDPPlanners.core.tree.anytree_based import (
-        ActionNode, BeliefNode, get_optimal_action_cost_setting,
-    )
-"""
-
-from POMDPPlanners.core.tree.anytree_based.base_node import BaseNode
-from POMDPPlanners.core.tree.anytree_based.action_node import ActionNode
-from POMDPPlanners.core.tree.anytree_based.belief_node import BeliefNode
-from POMDPPlanners.core.tree.anytree_based.visualization import print_tree
-from POMDPPlanners.core.tree.anytree_based.optimal_action import (
-    get_optimal_action_cost_setting,
-    get_optimal_action_reward_setting,
-)
-
-__all__ = [
-    "BaseNode",
-    "ActionNode",
-    "BeliefNode",
-    "print_tree",
-    "get_optimal_action_cost_setting",
-    "get_optimal_action_reward_setting",
-]
